demo/requirements.txt
openai>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
pulp>=2.8.0
python-dotenv>=1.0.0
